lcov: make 0%/100% exclusive to no/full coverage rate
Ensure that coverage rates 0% and 100% are only used when no or all
lines/functions/branches are hit respectively. This approach is
implemented to allow better identification of boundary cases, and
to be in accordance with the behavior of the gcov tool.

Based on suggestions by: Paul.Zimmermann@loria.fr and
vincent@vinc17.net