genhtml: improve function table sorting
In the function table view, the initial view should show the functions
sorted by execution count because - unlike with file names - the function
name is not a natural order for functions (the line number would be,
but that is not available). Also sort functions with the same execution
count alphabetically for a stable order.

Base on a suggestion by paul.bignier@hotmail.fr.