geninfo: Fix error when using --demangle-cpp
Using genhtml's --demangle-cpp option on data produced with recent GCC
versions (at least 4.8 and 4.9) can result in an error message similar
to the following:

  genhtml: ERROR: Demangled function name _ZN3subD2Ev  maps to different
           lines (5 vs 4)

The reason for this error is an unexpected sequence of lines records
in a .gcno file. These records mention line numbers as belonging to a
function which occur before the initial line number of that function
as reported by the corresponding function record.

Fix this problem by retaining the order of lines belonging to a function
as found in the .gcno file. This way geninfo will consistently use the
initial line number as reported by the function record when merging
function data during the demangling process.

Reported-by: Alexandre Duret-Lutz <adl@lrde.epita.fr>
Signed-off-by: Peter Oberparleiter <oberpar@linux.vnet.ibm.com>