lcov: Enable userspace package capture with only data files
Previously lcov's --from-package capture mechanism required
that .gcno files and source were present on the test machine.

This patch modifies --from-package capturing to work when
only .gcda files are present in the package captured on the
test machine. It works by linking the .gcda files collected
on the test machine into their natural location on the build
machine. This requires existing .gcda files to be removed.

Signed-off-by: Peter Oberparleiter <oberpar@linux.vnet.ibm.com>