geninfo: Fix 'unexpected end of file' error
Use the compiler version as stored in the .gcno file to determine if
the file contains function records with split checksums. This fixes
the following problem that can occur when lcov is run using a gcov
tool of GCC version 4.7 and above on .gcno files compiled with a
version below 4.7:

  # lcov -c -d . -o test.info --initial
  [...]
  geninfo: ERROR: test.gcno: reached unexpected end of file

Also add missing lcov version to --debug output.

Signed-off-by: Peter Oberparleiter <oberpar@linux.vnet.ibm.com>